requirements.txt
torch
transformers
diffusers
requests
Pillow
torchaudio  
librosa     
ollama
